Refactored code, modular character.py now
